IO Reader: Avoid clearing the scratch buffer before writing
The IoRead::read implementation used to manage the length of the vector
and its content independently, necessitating a zero-initialization. In
this change, only the capacity of the vec is set in advance, and the vec
is fed the input data from a taken read slice's read_to_end method,
which populates the vector and updates its length in (optimized)
lockstep.

As a result, the buffer has the nice property of always being the right
length, so scratch_offset has become a bit redundant -- but not yet been
removed, because there's two directions to go: Either

* do away with scratch_offset (as the information is already transfered
  in the length of the vec), or

* to move the buf (whose management is currently shared responsibility
  between deserializer and reader) completely into the reader, so
  `.read(n, scratch, scratch_offset)` becomes `.clear_buffer();
  .read_into_buffer(n)`, and the fully populated buffer can then be
  fetched as `.read_buffer<'a>() -> &'a [u8]`.

  Such a scheme would be helpful in implementations for no_std
  environments (either without alloc or with fallible allocation), and
  help reduce the number of #[cfg()] around the deserializer's infinite
  string handling there.