de/Read: Move scratch buffers to readers
This simplifies the parse_indefinite_* functions as they can now rely on
the full string to be in the reader's scratch buffer.

It adds a bit of interface to the Reader as the deserializer can now
specifically ask for reads to go directly to the scratch buffer or to
have either way (scratch or reference), plus the new function for
clearing and accessing the buffer -- but then again the functions
themselves are more straightforward (no passing of a buf that may or may
not be ignored by implementations).