IO Reader: Don't loop in 16k chunks
The splitting of reads into 16k chunks to avoid resource exhaustion has
become obsolete since .take() / .read_to_end() is used, as its
implementation already takes care of allocating in sane chunk sizes.

The .reserve() call for the small values ensures that we use the
available knowledge in the typical cases.

(Given indenting changes, this commit is best viewed with
--ignore-all-space).